requirements.txt
# MedAI - Medical Transcription & Analysis System
# Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML
openai-whisper==20231117
google-generativeai==0.3.2
torch>=2.0.0

# Image Processing
Pillow==10.2.0

# PDF Generation
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
